Codestyle

+++ api/src/main/java/com/velocitypowered/api/proxy/server/RegisteredServer.java
@@ -36,13 +36,12 @@ public interface RegisteredServer extends ChannelMessageSink, Audience {
    */
   Collection<Player> getPlayersConnected();
 
-
   /**
    * Get the total player count of the server (redis support).
    *
    * @return The total player count.
    */
-  int getTotalPlayerCount();
+  long getTotalPlayerCount();
 
   /**
    * Returns a list of all the players currently connected to this server on all proxies
